underwater-detection-gradio/requirements.txt
gradio
ultralytics
opencv-python
Pillow
torch>=2.0.0 
torchvision>=0.15.0